refactor: replace differ with diff-then-lower composition
